Implemented Functioning Shooter Class.
Created the Shooter class. It is fully implemented to run on joystick 4
of 4. The shooter talks to CANTalons 5 and 7. Since there was conflict
with those CANTalon channels being used by another aspect of the code,
conflicts were commented out.

Deleted the RobotOutputs java file, there is no need for the that class
since it does not follow the structure of our newly agreed code design.

Modified RobotCoDriver to call the shooter class and pass the joystick
variable so that the shooter can access the joystick.