Abstracted the shooter to reduce large classes
Created a FlyWheel class that takes care of all handling regarding the
flywheel. This was done to reduce classes with lots of code. Redesigned
the shooter class to not use the joystick and use the module ArrayList
in the Robot class. For shooter testing it was connected to the
SmartDashboard for setting values and reading the speed.

Added RPS (revolotions per second) calculation to the FlyWheel class.

Removed unused code in RobotCoDriver