Added different Driving modes.
Modified the DriveSide class such that there can be a powerLimit
variable that limits the power output to the motors.
Modified the DriveSide class such that there can be a cubeCurve to
powering the motor.
Modified the TankDriveCalculator such that it drives the robot with two
sticks like a tank.
Modified RobotDriver so that we can choose between different driving
modes.
Modified Vectors such that there can be inverted throttle and such that
we can drive using thankdrive control.