Add velocity control for flywheel, Added a new CANTalon class
The new Flywheel class should be able to use a PID loop, has not been
tested. Zack needs to use it.
The new CANTalon class works in a way that lets the CANTalon as a source
such that it can use encoder velocity as the error when calling the
pidGet method.