Fixed Command Problem with Pete's Button Fix
Pete's solution for having the buttons do things even when joystick is
in deadband made it such that the robot was always stopped regardless of
weather there is a command running. The problem was that the robot was
stopped everytime usingJoystick() was called and that was called every
iteration of the loop.