Added Arm Springyness to code
The arm would backdrive due to gravity. So we added a 0.1 idle power.
hat would move the arm back when it is at 90 degrees. So we used the cos
of the angle multiplied with the idle power so that it adds idle power
when it is getting more gravity force and less when not.