Added several buttons and extra Camera.
Added a third camera functionality for the Computer Vision class.
Added an Invert Button(#9) that also switches to back camera.
Added an uninvert Button(#8) that also switches to front camera.
Added an auto routine button(#8) for codriver that will perform the
selected auto routine.
Added a top spin button(#9) shot that will spin the flywheel and move
the arm to an angle.