Mid Competition Changes. We just lost :(
A significant change is that the auto aim command now also requires the
arm. Also when the driver starts moving the drive train power command is
called once to stop any other commands using the drive train. For the co
driver the same applies. This is to not have happen what happened in a
match where auto aim was going crazy because of exposure setting
problems and the robot froze mid match. Also when switching camera feed
the command re sets the exposure every switch to insure that the
exposure for that camera stays where we need it. A problem came when the
driver camera broke.
Added a camera switch button in co driver first joystick thumb buttons.
Adjusted how the offset's are being set so now it is a universal offset
and the constructor changes the default offset by the given number.
Added a servo actuator class to move the long 15 inch extenders to be
able to not be blocked when shooting high.
Rezeroed the arm. Changed some preferences.