Tried to fix flywheel orientation.
Flywheel was spinning wrong direction, and when I changed the inversion
boolean it started dropping suddenly back to zero. When I ran power, it
worked fine and the encoder was reading the correct values, but the PID
was still mysteriously screwed up in only one direction. Not Sure what's
going on here.